ETSWORK-16638 add hooks for query and modal state
