+++ examples/react/json-based-notification-renderer-with-custom-components/src/components/Notifications.jsx
@@ -14,77 +14,42 @@
  * limitations under the License.
  */
 
-import React, { useEffect, useState } from 'react'
-import { useQuery, gql } from '@apollo/client'
+import React from 'react'
 import Notification from './Notification'
 import Modal from './Modal'
 import Box from '@mui/material/Box'
 import LinearProgress from '@mui/material/LinearProgress'
 import { TransitionGroup } from 'react-transition-group'
 import Grow from '@mui/material/Grow'
-
-// GraphQL Query for list of notifications
-const GET_NOTIFICATIONS_QUERY = gql`
-  query GetNotifications($spaceId: ID!, $in: [NotificationLocation!]!) {
-    me {
-      space(spaceId: $spaceId) {
-        notifications(in: $in) {
-          edges {
-            node {
-              id
-              isDeletable
-              view
-              location
-            }
-          }
-        }
-      }
-    }
-  }
-`
+import useFetchNotifications from './hooks/useFetchNotifications'
+import useDeleteNotification from './hooks/useDeleteNotification'
+import useModal from './hooks/useModal'
 
 export default function Notifications() {
-  const [openModal, setOpenModal] = useState(false)
-  const [payload, setPayload] = useState(undefined)
-  const [notifications, setNotifications] = useState([])
+  const [payload, isOpen, handleOpenModal, handleCloseModal] = useModal(false)
 
-  const handleOpenModal = (cardPayload) => {
-    setOpenModal(true)
-    setPayload(cardPayload)
-  }
+  const [notifications, setNotifications, loading, error] = useFetchNotifications()
 
-  const handleCloseModal = () => {
-    setOpenModal(false)
-    setPayload(undefined)
-  }
+  const [deleteNotification] = useDeleteNotification()
 
   const handleDelete = (id) => {
+    deleteNotification({ variables: { input: { spaceId: process.env.REACT_APP_SPACE_ID, notificationId: id } } })
     setNotifications(notifications.filter((notification) => notification.node.id !== id))
   }
 
-  // GraphQL Query hook to automatically fetch data
-  const { data, loading, error } = useQuery(GET_NOTIFICATIONS_QUERY, {
-    variables: { spaceId: process.env.REACT_APP_SPACE_ID, in: ['TOKNOW', 'TODO'] },
-  })
-
-  const notificationEdges = data?.me?.space?.notifications?.edges
-
-  useEffect(() => {
-    notificationEdges && setNotifications(notificationEdges)
-  }, [notificationEdges])
-
-  // Content is not ready, show error/loader
-  if (loading)
+  if (loading) {
     return (
       <Box>
         <LinearProgress />
       </Box>
     )
+  }
+
   if (error) return <pre>{error.message}</pre>
 
   return (
     <div>
-      <Modal open={openModal} payload={payload} handleClose={handleCloseModal} />
+      <Modal open={isOpen} payload={payload} handleClose={handleCloseModal} />
       <TransitionGroup>
         {notifications.map((edge, idx) => (
           <Grow
