+++ examples/react/json-based-notification-renderer-with-custom-components/src/components/hooks/useModal.jsx
@@ -1,0 +1,17 @@
+import { useState } from 'react'
+
+export default function useModal() {
+  const [isOpen, setIsOpen] = useState(false)
+  const [payload, setPayload] = useState(undefined)
+
+  const handleOpenModal = (cardPayload) => {
+    setIsOpen(true)
+    setPayload(cardPayload)
+  }
+
+  const handleCloseModal = () => {
+    setIsOpen(false)
+    setPayload(undefined)
+  }
+  return [payload, isOpen, handleOpenModal, handleCloseModal]
+}
